requirements.txt
asgiref==3.8.1
black==24.4.2
click==8.1.7
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
sqlparse==0.5.0
